refactor: 참가자 추상 클래스로 추상화
