+++ src/main/java/domain/participant/Dealer.java
@@ -2,16 +2,13 @@ package domain.participant;
 
 import domain.card.Card;
 import domain.card.CardDeck;
-import domain.card.Cards;
-import java.util.List;
 
-public class Dealer {
+public class Dealer extends Participant {
     private final CardDeck cardDeck;
-    private final Cards ownedCards;
 
-    private Dealer(CardDeck cardDeck) {
+    public Dealer(CardDeck cardDeck) {
+        super();
         this.cardDeck = cardDeck;
-        this.ownedCards = Cards.of();
     }
 
     public static Dealer of(CardDeck cardDeck) {
@@ -20,21 +17,5 @@ public class Dealer {
 
     public Card drawCard() {
         return cardDeck.popCard();
-    }
-
-    public void receive() {
-        ownedCards.add(cardDeck.popCard());
-    }
-
-    public int getScore() {
-        return ownedCards.calculateScore();
-    }
-
-    public int getCardCount() {
-        return ownedCards.getSize();
-    }
-
-    public List<Card> getOwnedCards() {
-        return ownedCards.getCards();
     }
 }
