+++ src/main/java/domain/participant/Player.java
@@ -1,38 +1,18 @@
 package domain.participant;
 
-import domain.card.Card;
-import domain.card.Cards;
 import domain.result.BlackjackResult;
-import java.util.List;
 
-public class Player {
+public class Player extends Participant {
 
     private final String name;
-    private final Cards ownedCards;
 
     private Player(final String name) {
+        super();
         this.name = name;
-        this.ownedCards = Cards.of();
     }
 
     public static Player of(final String name) {
         return new Player(name);
-    }
-
-    public void receive(final Card card) {
-        ownedCards.add(card);
-    }
-
-    public int getScore() {
-        return ownedCards.calculateScore();
-    }
-
-    public int getCardCount() {
-        return ownedCards.getSize();
-    }
-
-    public List<Card> getOwnedCards() {
-        return ownedCards.getCards();
     }
 
     public String getName() {
