+++ src/test/java/domain/result/BlackjackResultTest.java
@@ -21,8 +21,8 @@ class BlackjackResultTest {
                 Card.of(TrumpNumber.SIX, TrumpShape.CLUB)
         );
         Dealer dealer = Dealer.of(CardDeck.of(cards));
-        dealer.receive();
-        dealer.receive();
+        dealer.receive(dealer.drawCard());
+        dealer.receive(dealer.drawCard());
         Player player = Player.of("pobi1");
         player.receive(Card.of(TrumpNumber.ACE, TrumpShape.CLUB));
         player.receive(Card.of(TrumpNumber.SEVEN, TrumpShape.CLUB));
